refactor(runtime): Move test-only logic from runner to test harness
Move _execute_on_device, validate_golden, build_orch_args_from_inputs,
golden_writer.py, and tensor_spec.py from pypto.runtime to the test
infrastructure (tests/st/harness/core/). Remove TensorSpec/ScalarSpec
from the public runtime API and golden_data_dir from RunConfig.

The core runtime now provides only compile-and-execute functionality,
while data generation and validation are test-level concerns.

Fixes #1043

Made-with: Cursor
