+++ tests/st/harness/core/golden_writer.py
