+++ tests/st/harness/core/tensor_spec.py
